feature(data.initdb): Added secondary index
My plan here is to store all the datafields together in the same table
and retrieve them by timestamp, though I wonder if this is even useful.
The data is in such a specific form that having it through time is
probably not super helpful.  Having the raw data through time is
considerably more so.

None